+++ lib/buildPackage.nix
@@ -4,6 +4,7 @@
 , lib
 , mkCargoDerivation
 , removeReferencesToVendoredSourcesHook
+, removeReferencesToRustToolchainHook
 , vendorCargoDeps
 }:
 
@@ -88,6 +89,7 @@ mkCargoDerivation (cleanedArgs // memoizedArgs // {
     # NB: avoid adding any non-hook packages here. Doing so will end up
     # changing PKG_CONFIG_PATH and cause rebuilds of `*-sys` crates.
     installFromCargoBuildLogHook
+    removeReferencesToRustToolchainHook
     removeReferencesToVendoredSourcesHook
   ];
 })