+++ lib/setupHooks/removeReferencesToRustToolchain.nix
@@ -1,0 +1,9 @@
+{ makeSetupHook
+}:
+makeSetupHook
+{
+  name = "removeReferencesToRustToolchainHook";
+  substitutions = {
+    storeDir = builtins.storeDir;
+  };
+} ./removeReferencesToRustToolchainHook.sh